tools/scrape_talks/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
